+++ cypress/integration/accessibility.spec.js
@@ -19,8 +19,8 @@ describe("Accessibility (A11Y) Check", () => {
     };
     cy.request("/api/templates", body)
       .then((response) => {
-        response.body.data.records.forEach((rec) => {
-          if (rec.formConfig.publishingStatus) {
+        response.body.data.forEach((rec) => {
+          if (rec.publishingStatus) {
             forms.push(rec.formID);
           }
         });
@@ -31,5 +31,11 @@ describe("Accessibility (A11Y) Check", () => {
         cy.injectAxe();
         cy.checkA11y(null, A11Y_OPTIONS);
       });
+  });
+
+  it("Check error state accessibility", () => {
+    cy.mockForm("../../tests/data/cdsIntakeTestForm.json");
+    cy.injectAxe();
+    cy.checkA11y(null, A11Y_OPTIONS);
   });
 });