+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- ReCAPTCHA V3 added on form submission
+
 ### Fixed
 
 - Aligned HTTP methods on API requests to decommission request body `method` property.
@@ -14,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Open links from richText component in new Tab.
 - Improved/fixed accessibility for file input component
 - Fixed dropdown initial value not being displayed the same way across browsers
+- Only public form template properties are available to unauthenticated sessions
 
 ## [1.1.0] 2022-03-04